Add notes, move an authentication function, fix a missing setSettings import
